moved mealApi.ts out of api folder since vercel sees files in the api folder differently and all frontend code should be out of it
