+++ app/components/MealPicker.tsx
@@ -1,5 +1,5 @@
 // components/MealPicker.tsx
-import { postMeal } from '@/api/mealApi';
+import { postMeal } from '@/util/mealApi';
 import { Picker } from '@react-native-picker/picker';
 import React, { useState } from 'react';
 import { Alert, Modal, Pressable, Text, View } from 'react-native';
